fix(types): avoid compiler error when dynamic dimensions are given by resolving type to `any[]`
When a Matrix is created without knowing it's exact size or number of dimensions, such as when
creating one dynamically, it's impossible recursively construct a type for it. We now short circuit
the type to resolve to `any[]`, to stop the typescript compiler throwing an `type instantiation is
excessively deep...` error. This ensures an array of at least one dimension is returned, but can't
provide more safety than that.